+++ server/dao/github.dao.ts
@@ -1,7 +1,12 @@
-import { getApiConfiguration } from "@/server/config/api/github.config";
-import { GitHubDeletionStatusType, type GitHubDeployments,  type GitHubWorkflowRun, type GitHubWorkflowRunApiResponse } from "@/server/types/github.type";
+import { getApiConfiguration } from "@@/server/config/api/github.config";
+import {
+  GitHubDeletionStatusType,
+  type GitHubDeployments,
+  type GitHubWorkflowRun,
+  type GitHubWorkflowRunApiResponse,
+} from "@@/server/types/github.type";
 
-import { logger } from "@/server/utils/logger";
+import { logger } from "@@/server/utils/logger";
 
 const api = getApiConfiguration();
 
@@ -53,9 +58,8 @@ export async function getWorkflowRuns(
   account: string,
   repository: string,
   token: string,
-  page: number = 1
+  page: number = 1,
 ): Promise<GitHubWorkflowRunApiResponse> {
-  // @ts-ignore
   return await api(`/repos/${account}/${repository}/actions/runs`, {
     method: "GET",
     headers: {
@@ -93,9 +97,9 @@ export async function getWorkflowRuns(
  * @returns A promise that resolves to an array of GitHubDeployments.
  * @throws An error if there is an issue retrieving the deployments.
  */
-export async function getDeployments(account: string, repository: string, token: string, page: number = 1): Promise<GitHubDeployments[]> {
+export async function getDeployments(account: string, repository: string, token: string, _page: number = 1): Promise<GitHubDeployments[]> {
   try {
-    return await api(`/repos/${account}/${repository}/deployments`, {
+    return await api<GitHubDeployments[]>(`/repos/${account}/${repository}/deployments`, {
       method: "GET",
       headers: {
         Authorization: `Bearer ${token}`,
@@ -130,13 +134,13 @@ export async function deleteWorkflowRun(account: string, repository: string, tok
       headers: {
         Authorization: `Bearer ${token}`,
       },
-      async onResponse({ request, response, options }) {
+      async onResponse({ response }) {
         if (response.status === 204) {
           logger.debug(`Workflow run [${id}] from [${account}/${repository}] repository deleted successfully`);
           deletionStatus = GitHubDeletionStatusType.SUCCESS;
         }
       },
-      async onResponseError({ request, response, options }) {
+      async onResponseError({ response }) {
         switch (response.status) {
           case 404:
             logger.debug(`Workflow run [${id}] from [${account}/${repository}] repository not found`);